requirements.txt
mysql-connector-python==8.2.0
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
requests==2.31.0
paho-mqtt==2.1.0
